refactor: Move more name resolution related functions to resolve.rs
